some refactoring; prepare inotify
