+++ types/types.go
@@ -1,5 +1,9 @@
 package types
 
+import (
+	"database/sql"
+)
+
 type FTypeStat struct {
 	NumBytes  uint64
 	FileCount uint
@@ -15,3 +19,15 @@ type FTypeDirStat struct {
 }
 
 type FileTypeDirStats map[string]*FTypeDirStat
+
+type FileTypeStatsDB struct {
+	// self *FileTypeStatsDB
+	fileName string
+	DB       *sql.DB
+}
+
+type DirTreeInfo struct {
+	statsDB   *FileTypeStatsDB
+	watchDirs []string
+	// notifyWatcher *inotify_blabla
+}
